lift & shift & adapt component formerly located in transactions.rs module.
